+++ R/ContainR_package.R
@@ -1,0 +1,15 @@
+#' aggreCAT: mathematically aggregating expert judgments
+#'
+#' To learn more about \pkg{ContainR}, start with the vignettes: \code{browseVignettes(package = "ContainR")}
+#'
+#' @docType package
+#' @name ContainR
+#'
+#' @keywords internal
+"_PACKAGE"
+
+# The following block is used by usethis to automatically manage
+# roxygen namespace tags. Modify with care!
+## usethis namespace: start
+## usethis namespace: end
+NULL