+++ R/data.R
@@ -1,0 +1,14 @@
+#' data_rocker_table
+#'
+#' A data table of rocker versioned images
+#'
+#' @format A data frame:
+#' \describe{
+#'  \item{name}{name of rocker image}
+#'  \item{image}{image of rokcer stack}
+#'  \item{base_image}{what the image is build on}
+#'  \item{description}{Short description}
+#' }
+#'
+#' @source \url{https://rocker-project.org/images/}
+"data_rocker_table"