+++ NEWS.md
@@ -1,3 +1,7 @@
 # ContainR 0.0.1
 
+* Added script to generate a Dockerfile based on local installed or loaded libs.
+* Added base functionality to docker commands.
 * Added a `NEWS.md` file to track changes to the package.
+* Initial work in progress release.
+